requirements.txt
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
apache-tvm-ffi==0.1.8.post2
asgiref==3.11.1
astor==0.8.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
black==26.1.0
blake3==1.0.8
blis==1.3.3
cachetools==7.0.0
catalogue==2.0.10
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
compressed-tensors==0.13.0
confection==0.1.5
contourpy==1.3.3
coverage==7.13.3
cryptography==46.0.4
csvw==3.7.0
ctranslate2==4.7.1
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cv2_enumerate_cameras==1.3.3
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
decorator==5.2.1
depyf==0.20.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
easydict==1.13
eclipse-zenoh==1.7.2
einops==0.8.2
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
fastapi==0.128.4
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
faster-whisper==1.2.1
fastrlock==0.8.3
filelock==3.20.3
flake8==7.3.0
flashinfer-python==0.6.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gguf==0.17.1
grpcio==1.78.0
grpcio-reflection==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ijson==3.4.0.post0
ImageIO==2.37.2
importlib_resources==6.5.2
iniconfig==2.3.0
insightface==0.7.3
interegular==0.3.3
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kokoro==0.9.4
language-tags==1.2.0
lark==1.2.2
lazy_loader==0.4
librosa==0.11.0
librt==0.7.8
libusb-package==1.0.26.3
livekit==1.0.25
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
log-throttling==0.0.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
misaki==0.9.4
mistral_common==1.9.0
ml_dtypes==0.5.4
model-hosting-container-standards==0.1.13
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
ninja==1.13.0
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.5
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnx-asr==0.10.2
onnxruntime==1.24.1
onnxruntime-gpu==1.24.1
openai==2.17.0
openai-harmony==0.0.8
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openwakeword==0.4.0
orpheus-speech==0.1.0
outlines_core==0.2.11
packaging==26.0
partial-json-parser==0.2.1.1.post7
pathspec==1.0.4
phonemizer-fork==3.3.2
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.9.0
preshed==3.0.12
prettytable==3.17.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
PyAudio==0.2.14
pybase64==1.4.3
pycodestyle==2.14.0
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
pyserial==3.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytokens==0.4.1
pyusb==1.3.1
PyYAML==6.0.3
pyzmq==27.1.0
questionary==2.1.1
ray==2.53.0
rdflib==7.5.0
reachy_mini==1.3.0
reachy_mini_motor_controller==1.5.4
reachy_mini_rust_kinematics==1.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
Resemblyzer==0.1.4
rfc3986==1.5.0
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rpds-py==0.30.0
rustypot==1.4.2
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
segments==2.3.0
sentencepiece==0.2.1
sentry-sdk==2.52.0
setproctitle==1.3.7
setuptools==80.10.2
shellingham==1.5.4
silero-vad==6.2.0
simsimd==6.5.12
six==1.17.0
smart_open==7.5.0
snac==1.2.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speexdsp==0.1.1
srsly==2.5.2
sse-starlette==3.2.0
starlette==0.52.1
stringzilla==4.6.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.1.28
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.3
transformers==4.57.6
triton==3.5.1
typer==0.21.1
typer-slim==0.21.1
types-protobuf==6.32.1.20251210
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.15.1
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
webrtcvad==2.0.10
websockets==15.0.1
wrapt==2.1.1
xgrammar==0.1.29
yarl==1.22.0
